Streamlined usage of Sprintf instead of strconv.
Streamlined usage of url.Values instead of string concatenation. This
avoids encoding issues and many cases where "&" and "?" were intermixed,
or missing.

Streamlined usage of named returns, for idiomatic 3-6 liners..

Implemented a generic "loadOptionalFilters" instead of the
"applyFiltersFor[this-and-that]" functions. (IntList is part of that)

Opened "httpClient" for people to use their own.

StatusCode > 299 could be a redirect, and wouldn't be an error. >399 are
errors.

btoitos -> boolToString

AssignedtoID -> AssignedToID (for idiomatic Go)

Offest -> Offset

TODO: this thing needs more tests.. (low hanging fruit: loadOptionalValues)